+++ man-roxygen/consensus-common-args.R
@@ -1,3 +1,9 @@
+#' @param absMinAbundance,relMinAbundance Minimum absolute or relative
+#'   (\samp{0-1}) abundance across objects for a result to be kept. For
+#'   instance, \code{relMinAbundance=0.5} means that a result should be present
+#'   in at least half of the number of compared objects. Set to \samp{NULL} to
+#'   ignore and keep all results. Limits cannot be set when \code{uniqueFrom} is
+#'   not \code{NULL}.
 #' @param uniqueFrom Set this argument to only retain <%=what%> that are unique
 #'   within one or more of the objects for which the consensus is made.
 #'   Selection is done by setting the value of \code{uniqueFrom} to a